third_party/WebKit/LayoutTests/fast/html/imports/custom-element-style-block.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel="import" href="resources/custom-element-style.html">
</head>
<body>
<x-test></x-test>
<script>
test(function(){
    var x = document.querySelector('x-test');
    var port = x.shadowRoot.querySelector('style').sheet.cssRules[0];
    assert_true(port.styleSheet instanceof StyleSheet);
}, "The @import style in custom element shadow DOM blocks following script element.");
</script>
</body>
</html>
